refactor(cozeloop-ai): mock -> __mock__
