+++ packages/cozeloop-ai/vitest.config.ts
@@ -16,7 +16,9 @@ export default defineConfig({
       ],
     },
     env: {
-      LOOP_VERSION: packageJson.version,
+      COZELOOP_VERSION: packageJson.version,
+      COZELOOP_WORKSPACE_ID: '7308703665823416358',
+      COZELOOP_API_TOKEN: 'pat_xxx',
     },
   },
 });
